fixed verification and added keep me signed in checkbox.
verification introduced a new data field inputValid that refreshes upon every submission. The error is still removed every time the user attempts
to enter some new text in either of the input fields.

checkbox was introduced to allow the backend to send a refresh token with a longer expiration date if the user says that they wish to stay signed
in.